refactor(engineering): 把工程实战工作台改成可按项目组合
原先「任务描述 + IDE + 结果面板」这套布局写死在 pages/projects/[id].tsx 里。
要加入内网设施那类需要终端和拓扑图的项目，工作台得能按项目形态组合。

- 项目定义新增可选的 workspace.kind（'code' | 'ops'），不写等同于 'code'，
  现有 9 个项目一个字段都不用加，走的也仍然是原来那条代码路径
- 会话逻辑（载入、进度、草稿、防抖落盘、切关卡、切语言、重置）原样搬进
  useProjectSession，与工作台形态无关，各形态共用
- 原有布局原样搬进 CodeWorkspace，页面退化成 30 行的分发器
- 形态未知时给出可读提示，而不是白屏

跑验收后记一笔尝试这件事，从页面内联的 persist 提成会话上的 recordAttempt；
「清结果」区分 'all' 与 'report' 两种范围，保住「重置本关」不清评分卡这个既有差别。
